+++ src/components/Comment/Comment.module.scss
@@ -13,12 +13,7 @@
     width: 100%;
     padding: 3rem;
     row-gap: 0.85em;
-    .commentList {
-      display: grid;
-      justify-items: flex-end;
-      width: 100%;
-      row-gap: 0.85em;
-    }
+
     .gridItem {
       position: relative;
       width: 100%;
@@ -79,6 +74,25 @@
       }
       .commentContent:disabled {
         background-color: #fff;
+      }
+    }
+    .pagenation {
+      display: flex;
+      gap: 1rem;
+      overflow: hidden;
+      .moveBtn {
+        z-index: 999;
+      }
+      .pageList {
+        display: flex;
+        gap: 2rem;
+        width: 5.4rem;
+
+        .pageItem {
+          width: 1rem;
+          height: 1rem;
+          cursor: pointer;
+        }
       }
     }
   }
