+++ src/components/Comment/CommentList/CommentList.module.scss
@@ -1,0 +1,6 @@
+.commentList {
+  display: grid;
+  justify-items: flex-end;
+  width: 100%;
+  row-gap: 0.85em;
+}
