--- src/components/Comment/Reply/Paging/Paging.scss
@@ -1,18 +1,0 @@
-.pagination {
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-  & li {
-    width: 1rem;
-    text-align: center;
-  }
-  .disabled a {
-    color: #f4f4f4;
-  }
-  .active {
-    background-color: #ff6d6a;
-    & a {
-      color: #fff;
-    }
-  }
-}
