+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-
+import 'bootstrap/dist/css/bootstrap.min.css';
 import './styles/reset.scss';
 import './styles/common.scss';
 import Router from './Router';
